requirements.txt
pydicom>=2.3.0
Pillow>=9.0.0
requests>=2.28.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
numpy>=1.21.0
matplotlib>=3.5.0 